+++ CHANGELOG.md
@@ -5,6 +5,8 @@ Functionality:
   - Custom Titlebars can now be shown in the focus tree preview (togglable)
   - Custom Focus Overlays can now be shown in the focus tree preview (togglable)
   - Focus Inlay Windows can now be shown in the focus tree preview (togglable)
+  - Changed feature settings to boolean instead of free text
+  - Add a setting to link technology interface files so the preview can look up other folders for technology icons
 
 Bugfixes:
   - Fixed error when copy paste the .mod file location. Now with "" the file will be located.